requirements.txt
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
gensim==4.3.3
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
nltk==3.9.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smart-open==7.1.0
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2024.2
wrapt==1.17.0
